Refactored `celestia.types`
