+++ celestia/node_api/state.py
@@ -2,7 +2,8 @@ from typing import Callable
 
 from celestia._celestia import types  # noqa
 
-from celestia.types import TxConfig, Blob, Unpack
+from celestia.types import TxConfig, Unpack
+from celestia.types.blob import Blob
 from celestia.types.state import Balance, TXResponse, QueryUnbondingDelegationResponse, \
     QueryDelegationResponse, QueryRedelegationResponse
 from celestia.node_api.rpc.abc import Wrapper
